+++ NEWS.md
@@ -4,6 +4,8 @@
 * Added a `method_only` argument to `d_proportion()`, `d_proportion_diff()`, and
   `d_test_proportion_diff()` to allow returning method labels without additional
   description. (#1525)
+* Added `forest_header_above` and `forest_title` arguments to `g_forest()` to
+  control the position of the forest plot header and add a forest plot title. (#1500)
 * Updated `g_forest()` to support point estimates and confidence intervals
   stored in a single column. (#1499)
 * Added the `exclude_rows` argument to `g_forest()` to allow excluding selected